third_party/blink/web_tests/virtual/popover-hint-new-behavior-disabled/external/wpt/html/semantics/popovers/popover-remove-attribute-during-focusing-steps-expected.txt

This is a testharness.js-based test.
[FAIL] Remove popover attribute during the popover focusing steps
  assert_equals: two toggle events should be dispatched expected 2 but got 1
Harness: the test ran to completion.
